+++ src/Stack/ConfigCmd.hs
@@ -14,6 +14,7 @@ module Stack.ConfigCmd
        ,cfgCmdName) where
 
 import           Control.Applicative
+import           Control.Exception
 import           Control.Monad
 import           Control.Monad.Catch (throwM)
 import           Control.Monad.IO.Class
@@ -23,13 +24,15 @@ import qualified Data.HashMap.Strict as HMap
 import           Data.Monoid
 import           Data.Text (Text)
 import qualified Data.Text as T
+import           Data.Typeable
 import qualified Data.Yaml as Yaml
 import qualified Options.Applicative as OA
 import qualified Options.Applicative.Types as OA
 import           Path
+import           Path.IO
 import           Prelude -- Silence redundant import warnings
 import           Stack.BuildPlan
-import           Stack.Config (makeConcreteResolver, getStackYaml)
+import           Stack.Config (makeConcreteResolver, getProjectConfig)
 import           Stack.Types.Config
 import           Stack.Types.Resolver
 
@@ -46,6 +49,16 @@ data CommandScope
       --   typically at @~/.stack/config.yaml@.
     | CommandScopeProject
       -- ^ Apply changes to the project @stack.yaml@.
+    deriving (Show)
+
+data ConfigCmdError
+    = ExpectedLocalProject
+    deriving (Typeable)
+
+instance Exception ConfigCmdError
+instance Show ConfigCmdError where
+    show ExpectedLocalProject =
+        "Since --global was not used, expected to find a local project configuration.  However, none was found."
 
 configCmdSetScope :: ConfigCmdSet -> CommandScope
 configCmdSetScope (ConfigCmdSetResolver _) = CommandScopeProject
@@ -54,14 +67,21 @@ configCmdSetScope (ConfigCmdSetInstallGhc scope _) = scope
 
 cfgCmdSet
     :: (StackMiniM env m, HasConfig env, HasGHCVariant env)
-    => ConfigCmdSet -> m ()
-cfgCmdSet cmd = do
+    => GlobalOpts -> ConfigCmdSet -> m ()
+cfgCmdSet go cmd = do
+    conf <- view configL
     configFilePath <-
         liftM
             toFilePath
             (case configCmdSetScope cmd of
-                 CommandScopeProject -> getStackYaml
-                 CommandScopeGlobal -> view $ configL.to configUserConfigPath)
+                 CommandScopeProject -> do
+                     mstackYamlOption <- forM (globalStackYaml go) resolveFile'
+                     mstackYaml <- getProjectConfig mstackYamlOption
+                     case mstackYaml of
+                         Just stackYaml -> return stackYaml
+                         Nothing -> throwM ExpectedLocalProject
+                 CommandScopeGlobal -> return (configUserConfigPath conf))
+    liftIO $ print (configCmdSetScope cmd, configFilePath)
     -- We don't need to worry about checking for a valid yaml here
     (config :: Yaml.Object) <-
         liftIO (Yaml.decodeFileEither configFilePath) >>= either throwM return